feat: move levels into subdirectories
